Move location fetch into result component
This keeps the page rendered while only the paginated list is updated, while avoiding to refetch the country data on every page selection
