+++ app/assets/scripts/views/country/country.js
@@ -1,10 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { PropTypes as T } from 'prop-types';
 import { connect } from 'react-redux';
-import _ from 'lodash';
-import qs from 'qs';
 
-import { buildQS } from '../../utils/url';
 import { getCountryBbox } from '../../utils/countries';
 import { formatThousands } from '../../utils/format';
 import config from '../../config';
@@ -12,21 +9,11 @@ import { openDownloadModal } from '../../actions/action-creators';
 import { ParameterProvider } from '../../context/parameter-context';
 
 import Header, { LoadingHeader, ErrorHeader } from '../../components/header';
-import LoadingMessage from '../../components/loading-message';
-import InfoMessage from '../../components/info-message';
 import MapComponent from '../../components/map';
 import LocationsSource from '../../components/map/locations-source';
 import MeasurementsLayer from '../../components/map/measurements-layer';
 import Legend from '../../components/map/legend';
 import Results from './results';
-
-const defaultLocations = {
-  locationFetching: false,
-  locationFetched: false,
-  locationError: null,
-  locations: null,
-  locationMeta: null,
-};
 
 const defaultCountry = {
   countryFetching: false,
@@ -34,92 +21,20 @@ const defaultCountry = {
   country: null,
 };
 
-const PER_PAGE = 30;
-
-function getPage(query) {
-  if (query && query.page) {
-    let page = query.page;
-    page = isNaN(page) || page < 1 ? 1 : +page;
-    return page;
-  }
-  return 1;
-}
-
-function Country({ match, history, location, _openDownloadModal }) {
+function Country({ match, _openDownloadModal }) {
   const id = match.params.name;
-  const [
-    {
-      locationFetching,
-      locationFetched,
-      locationError,
-      locations,
-      locationMeta,
-    },
-    setLocations,
-  ] = useState(defaultLocations);
+
   const [{ countryFetching, countryError, country }, setCountry] = useState(
     defaultCountry
   );
-  const [page, setPage] = useState(1);
-  const [isMounted, setIsMounted] = useState(false);
-  useEffect(() => {
-    setIsMounted(true);
-    let query = qs.parse(location.search, {
-      ignoreQueryPrefix: true,
-    });
-    setPage(() => getPage(query));
-  }, []);
 
   useEffect(() => {
-    if (!isMounted) return;
-    fetchLocations(id, page);
     fetchCountry(id);
 
     return () => {
-      setLocations(defaultLocations);
       setCountry(defaultCountry);
     };
-  }, [id, page, isMounted]);
-
-  const fetchLocations = (id, page) => {
-    setLocations(state => ({
-      ...state,
-      locationFetching: true,
-      locationFetched: false,
-      locationError: null,
-    }));
-
-    fetch(
-      `${config.api}/locations?page=${page}&limit=${PER_PAGE}&metadata=true&country=${id}`
-    )
-      .then(response => {
-        if (response.status >= 400) {
-          throw new Error('Bad response');
-        }
-        return response.json();
-      })
-      .then(
-        json => {
-          setLocations(state => ({
-            ...state,
-            fetched: true,
-            locationFetching: false,
-            locationFetched: true,
-            locations: json.results,
-            locationMeta: json.meta,
-          }));
-        },
-        e => {
-          console.log('e', e);
-          setLocations(state => ({
-            ...state,
-            locationFetching: false,
-            locationFetched: false,
-            locationError: e,
-          }));
-        }
-      );
-  };
+  }, [id]);
 
   const fetchCountry = id => {
     setCountry(state => ({
@@ -153,22 +68,11 @@ function Country({ match, history, location, _openDownloadModal }) {
       );
   };
 
-  function handlePageClick(d) {
-    let query = qs.parse(id, location.search, { ignoreQueryPrefix: true });
-    const pageNumber = d.selected + 1;
-    query.page = pageNumber;
-    setPage(pageNumber);
-
-    history.push(`/countries/${id}?${buildQS(query)}`);
-  }
-
-  let locationGroups = _(locations).sortBy('city').groupBy('city').value();
-  const totalPages = locationMeta && Math.ceil(locationMeta.found / PER_PAGE);
   return (
     <section className="inpage" data-cy="country-page">
-      {countryFetching || locationFetching ? (
+      {countryFetching ? (
         <LoadingHeader />
-      ) : country && locations && !countryError ? (
+      ) : country && !countryError ? (
         <>
           <Header
             id="country"
@@ -222,23 +126,7 @@ function Country({ match, history, location, _openDownloadModal }) {
                 </div>
               </section>
 
-              {locationFetching ? (
-                <LoadingMessage />
-              ) : !locationError ? (
-                <Results
-                  fetched={locationFetched}
-                  fetching={locationFetching}
-                  error={locationError}
-                  locationGroups={locationGroups}
-                  id={id}
-                  totalPages={totalPages}
-                  page={page}
-                  openDownloadModal={_openDownloadModal}
-                  handlePageClick={handlePageClick}
-                />
-              ) : (
-                <InfoMessage standardMessage />
-              )}
+              <Results id={id} openDownloadModal={_openDownloadModal} />
             </div>
           </div>
         </>
@@ -251,8 +139,6 @@ function Country({ match, history, location, _openDownloadModal }) {
 
 Country.propTypes = {
   match: T.object,
-  location: T.object,
-  history: T.object,
   _openDownloadModal: T.func,
 };
